requirements.txt
apns2==0.7.2
asgiref==3.4.1
astroid==2.5.8
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
django-push-notifications==2.0.0
django-redis==4.12.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
gunicorn==20.0.4
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==21.3
Pillow==8.1.2
pluggy==0.13.1
psycopg2==2.8.6
py==1.11.0
py-vapid==1.8.2
pycparser==2.21
PyJWT==1.7.1
pylint==2.7.2
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.3
pywebpush==1.14.0
redis==4.1.1
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
wrapt==1.12.1